Backend/requirements.txt
Flask
Flask-Cors
pandas
openpyxl
PyJWT
werkzeug